requirements.txt
numpy==1.24.3
pandas==2.0.3
yfinance==0.2.31
scikit-learn==1.3.0
stable-baselines3==2.1.0
gym==0.26.2
tensorboard==2.14.0
torch==2.0.1
matplotlib==3.7.2
pyyaml==6.0.1
loguru==0.7.2
